Fix for patch 7.37

+++ src/Magus.Data/Models/Dota/Item.cs
@@ -5,7 +5,7 @@ public record Item : BaseSpell
     public IEnumerable<string>? ItemAliases { get; set; }
 
     public short ItemCost { get; set; } // if cost is 0, and item is ItemRecipe, item does NOT exist to buy, even if ItemPurchasable is set... figure this out
-    public byte ItemInitialCharges { get; set; }
+    public int ItemInitialCharges { get; set; }
     public byte ItemStockInitial { get; set; }
     public byte ItemStockMax { get; set; }
     public float ItemStockTime { get; set; }
